move functions into helpers
moving endForceSync and forceTransition into their respective helpers
